refactor: platform-dependant code
