+++ boltconn/src/cli/cert.rs
@@ -1,4 +1,4 @@
-use nix::unistd::{Gid, Uid};
+use crate::platform::UserInfo;
 use rcgen::{
     date_time_ymd, BasicConstraints, Certificate, CertificateParams, DistinguishedName, DnType,
     IsCa, KeyUsagePurpose,
@@ -9,7 +9,7 @@ use std::path::{Path, PathBuf};
 
 fn safe_write(path: PathBuf, content: &str) -> anyhow::Result<()> {
     fs::write(path.as_path(), content)?;
-    nix::unistd::chown(&path, Some(Uid::from(0)), Some(Gid::from(0)))?;
+    UserInfo::root().chown(&path)?;
     let perm = fs::Permissions::from_mode(0o600);
     fs::set_permissions(&path, perm)?;
     Ok(())
